pop()
You have a list named cl with the following elements: [2, 3, 1].

You call the pop(2) method on the list cl. The pop() method in Python is used to remove and return an element from a list at a specified index. In this case, you're specifying index 2, which corresponds to the third element in the list (Python uses zero-based indexing).

The pop(2) method removes the element at index 2, which is the number 1, from the list cl.

The pop(2) method also returns the value that was removed, which is the number 1.

The print() function is used to display the value returned by cl.pop(2). So, it will print 1.

After executing this code, the list cl will be modified to [2, 3], and the number 1 will be printed to the console.