list and for loop
This Python code will iterate over the list lis using a for loop. Within each iteration, it will unpack the sublists into the variables p and q. Then it will print the sum of p and q followed by an ampersand (&). Let's go through the steps:

p and q will take the values from the sublists in lis successively.
In the first iteration, p will be 8 and q will be 7. The sum will be 15. So, it will print 15&.
In the second iteration, p will be 6 and q will be 5. The sum will be 11. So, it will print 11&.
Therefore, the output of the code will be: 15&11&