copy function
Step 1: Initialize the list n with two elements [76, 24].
n = [76, 24]

Step 2: Create a copy of the list n and assign it to the variable p using the copy() method. Both p and n will initially contain the same elements.
p = n.copy()

At this point, p and n are both [76, 24].

Step 3: Use the pop() method on the list n without specifying an index, which by default removes and returns the last element of the list. In this case, it removes the last element, 24, from the list n. So, after this line of code, n contains only [76].
n.pop()

Now, n is [76].

Step 4: Print the contents of the variables p and n. This will display the values of both lists at this point.
print(p, n)

The output will be:
[76, 24] [76]

p still contains the original values [76, 24], while n has had one element removed and now contains [76].