Update 2.py
2. Execute the given function.
def LargeSmallSum(arr)
The function takes an integral arr which is of the size or length of its arguments. Return the sum of the second smallest element at odd position ‘arr’ and the second largest element at the even position.
Assumption
o	Every array element is unique.
o	Array is 0 indexed.
Note:
o	If the array is empty, return 0.
o	If array length is 3 or <3, return 0.
 
Example
Input:
Arr: 3 2 1 7 5 4
Output:
7
 
Explanation
o	The second largest element at the even position is 3.
o	The second smallest element at the odd position is 4.
o	The output is 7 (3 + 4).